Draw Starting Pieces on Board
This introduces a variants.js file that tracks the initila config and
will eventually track rules for variants on hexagonal chess. For this
project we're only using Glinski at this time.

The new function will draw the pieces to the appropriate hex tiles for
the given startup.

Next, the pieces must be instantiated and tracked in the board. This is
not yet implemented.