Create a Game Module, plus code cleanup
This creates a Game Module with a new game method that constructs the
board and pieces and takes responsibility of managing and evoking
Raphael.

This also adds in some spaces, semicolons, trailing commas, etc. that
were bothering me from earlier / I've been too lazy to run JSLint so
there will be more undoubtedly once the game is functional.